Try to fix weird cursor position in prompt
Changed how and where the color codes are inserted into the prompt so
that the cursor will stop getting screwed up when I'm using tmux, but it
doesn't seem to have worked. Still, I think the prompt set up is a
little nicer, now that it's not using PROMPT_COMMAND and we set colors
in fewer places.