Experimenting with a "writing mode"
I'm playing around with some tricks to try to emulate the experience of
using an editor like iA Writer that removes a lot of the metadata from
the window and centers your content in the window for you. It's kinda
working, but it's not quite there yet. I suppose I could also look for a
plugin. I believe several exist.