Fix the prompt
I had an error in my echo syntax that resulted in no path being
displayed unless you were in a Git repo. I replaced the echo with a
printf, although I'm not sure that compressing the prompt to a single
line is an improvement over the echo statements, which were less terse
and didn't include an unintelligible format string.

For the record, you have to separate each argument to echo with a space,
so to avoid spaces between the path and the Git prompt, you need `echo
-s`, which removes spaces between the arguments.