Load a local config file if one is present for fish
I often have computer-specific settings, or settings that shouldn't be
committed to a Git repo because they contain secrets. So to handle that,
check for ~/.config/fish/local.fish, and if one is present, source it.
The local.fish file will not be added to Git, so it's safe for secrets,
etc.