Switch to Solarized Dark
Rae pointed out some cool Solarized Dark keycaps, and now I'm thinking
about getting them, so I'm switching back to Solarized Dark as my color
scheme for a bit. This doesn't include the installation of Solarized for
Vim because it doesn't work with Plug. I'll need to figure out a
solution for that.