Remove <space> keybinding for folds
This is interfering with a lot of the bindings for LSP. There's a slight
delay after hitting space because Vim is waiting to see if any other
keypresses will follow to trigger the combinations for LSP, when none
happens, it toggles the fold. But I'm just fine using zo and zc.