Build full path to vim config dirs
For some reason, ~ in the paths didn't seem like it was working well in
the lua config, so instead I'm building the full path in the config
using vim.fn.expand for my .vimrc and using vim.fn.stdpath('data') for
the undo directory. Hopefully this quits creating weird directories all
over the place.