Open a new window in the CWD by default
I generally prefer to open a new split in the CWD, since I use one tab
per-project and it's not hard to navigate to other directories thanks to
fzf on the rare occasion that's what I want. Therefore, I remapped the
keybinding for new_window to new_window_with_cwd.