Copying over good bits from LazyVim
I've been looking into how the LazyVim plugin configures things and
copying the bits that I like. I've stopped sourcing the regular Vim
config in favor of putting all of my options in an options lua module;
mainly because some of the settings I want for a plugin-less Vim
installation might be at odds with plugins in my NeoVim setup.

I'm also switching to lazy.nvim for plugin management. The lazy loading
is nice for keeping the startup times down. I've replaced FZF with
Telescope because it does similar things, but is native to NeoVim so it
can take advantage of things like the dialog UI supported by NeoVim.