Modify the basic dev layout slightly
Allow the NeoVim frame to have a frame, why not. Instead of editing ".",
which has the undesirable effect of leaving the NetRW buffer open if I
use telescope to open a file and cause :wa to give errors, I just let it
launch nvim. I'm going to go ahead and add the alpha plugin for a
dashboard on startup similar to LazyVim's setup, which should make this
initial view a little more useful.